Update software versions in GitHub Actions and requirements.txt
In response to compatibility issues, the software versions in the GitHub Actions configuration and requirements.txt file have been updated. Specifically, the Git actions for Python and Go have been set to use specific versions instead of the latest ones. In requirements.txt, the version of Protobuf has been downgraded to a more stable version.