Update IP addresses, package versions and fix query method
The IP addresses in the configuration files for the API, client service, and edge service have been updated. Package versions in the requirements.txt file have also been updated. A line of code in the edge.py file has been changed to correct the way a database query is made.