Update Dockerfile and configurations, add MySQL connectivity script
This commit does several modifications including updating the Dockerfile for building the api and installing required tools. It also updates the hostnames in the config.yaml file for api, consul, and mysql. Furthermore, it includes a new script run.sh to check for MySQL connectivity before starting the application. Finally, the created Docker image has been included in the docker-compose.yml.