Add consul connection check to run.sh script
The existing run.sh script was improved for checking the connection to not only MySQL but also Consul. The environment variables for MySQL host and port have been removed in favor of hardcoded values. The script will now output proper checks and establish connection to both MySQL on port 3306 and Consul on port 8500 respectively before running the main program.