Add GetIPAddress function and utilize it in API
Implemented a new function, GetIPAddress, that retrieves the IP address based on the runtime Operating System. This function replaces the previous method of hardcoding the host IP in the ApiConfig struct. As a result, the "host" field has been removed from ApiConfig and all associated files. The IP is now dynamically obtained when initiating the API server. This solution provides greater flexibility when deploying the API across different platforms.