Update common name across service and config files
The common name previously hardcoded as 'ccc' has been updated to 'FlowFederate' in the service/client and service/edge files. Config files for API, client and edge services have also been updated to take this common name from the configuration instead of hardcoding it, improving flexibility and ease of configurability.