Update release workflow with docker build and push
The release workflow in GitHub actions is updated to include Docker image building and pushing. The changes allow the workflow to automatically build Docker images for both 'flow-federate-sc' and 'flow-federate-api' and push them to Docker Hub upon successful tests, if triggered by a 'push' event. The GitHub token for 'semantic-release' is also switched to a bot token.