fix: Updated GitHub Actions setup and push logic
The version of the setup-go GitHub Action in build workflow has been upgraded to v5.0.1. Moreover, the logic for checking the need to push has been changed from using the action's output to setting an environment variable. Apart from this, the conditions in the Docker login, build and push actions have been updated to use the aforementioned environment variable.