feat: Add RabbitMQ connection check and update docker-compose files
This commit introduces a connection check to RabbitMQ in the run.sh script. It also updates the docker-compose and docker-compose.prod files to remove hard-coded MySQL root password and makes some changes in the service structure in docker-compose.yml. They are now built in the context of the current directory using a Dockerfile. This should help make services more flexible and secure.