feat: Update data types of 'accuracy' and 'loss' in multiple files
This commit involves changing the data types of 'accuracy' and 'loss' from `float` to `string` in several files to accommodate arrays. This allows storing more detailed accuracy and loss information over time as arrays in the database. It also refines the building process of Docker containers in 'docker-compose.yml' and the related 'Makefile'.